search criteria working

+++ backend/src/main/java/com/rental/listingservice/dto/ListingSearchCriteria.java
@@ -1,10 +1,10 @@
 package com.rental.listingservice.dto;
 
-public record ListingSearchCriteria(Double minPrice, Double maxPrice, Integer minBedrooms, String city, String keyword, int page, int items) {
+public record ListingSearchCriteria(Integer minPrice, Integer maxPrice, Integer minBedrooms, String city, String keyword, int page, int items) {
     public static final int DEFAULT_PAGE = 0;
     public static final int DEFAULT_ITEMS_PER_PAGE=10;
 
-    public static ListingSearchCriteria of(Double minPrice, Double maxPrice, Integer minBedrooms, String city, String keyword, Integer page, Integer items) {
+    public static ListingSearchCriteria of(Integer minPrice, Integer maxPrice, Integer minBedrooms, String city, String keyword, Integer page, Integer items) {
         String validatedCity = (city == null || city.isBlank()) ? null : city.trim();
         String validatedKeyword =  (keyword == null || keyword.isBlank()) ? null : keyword.trim();
